CST435_gRPC/requirements.txt
grpcio
grpcio-tools
cryptography
langdetect
vaderSentiment
deep-translator

